feat: introduce `miden-faucet` crate
