+++ bin/server/.env
