+++ bin/server/Cargo.toml
@@ -1,0 +1,62 @@
+[package]
+authors.workspace      = true
+description            = "Token faucet application for Miden testnet"
+edition.workspace      = true
+homepage.workspace     = true
+keywords               = ["faucet", "miden", "node"]
+license.workspace      = true
+name                   = "server"
+readme                 = "README.md"
+repository.workspace   = true
+rust-version.workspace = true
+version.workspace      = true
+
+[lints]
+workspace = true
+
+[dependencies]
+miden-faucet = { workspace = true }
+
+# Miden dependencies.
+miden-client     = { features = ["sqlite", "tonic"], workspace = true }
+miden-node-utils = { workspace = true }
+
+# External dependencies.
+anyhow       = { workspace = true }
+async-trait  = { version = "0.1" }
+axum         = { features = ["tokio"], version = "0.8" }
+axum-extra   = { version = "0.10" }
+base64       = { version = "0.22" }
+clap         = { features = ["derive", "env", "string"], version = "4.5" }
+http         = { workspace = true }
+humantime    = { workspace = true }
+rand         = { features = ["thread_rng"], workspace = true }
+rand_chacha  = { version = "0.9" }
+serde        = { features = ["derive"], version = "1.0" }
+serde_json   = { version = "1.0" }
+sha3         = { version = "0.10" }
+thiserror    = { workspace = true }
+tokio        = { features = ["fs"], workspace = true }
+tokio-stream = { features = ["net"], workspace = true }
+tonic        = { features = ["tls-native-roots"], workspace = true }
+tower        = { workspace = true }
+tower-http   = { features = ["cors", "set-header", "trace"], workspace = true }
+tracing      = { workspace = true }
+url          = { workspace = true }
+
+[build-dependencies]
+# Required to inject build metadata.
+miden-node-utils = { features = ["vergen"], workspace = true }
+
+[dev-dependencies]
+fantoccini       = { version = "0.22" }
+miden-node-proto = { branch = "next", git = "https://github.com/0xMiden/miden-node" }
+miden-testing    = { branch = "next", features = ["async"], git = "https://github.com/0xMiden/miden-base" }
+serde_json       = { version = "1.0" }
+tokio            = { features = ["process"], workspace = true }
+tonic-web        = { version = "0.13" }
+
+# Required to avoid false positives in cargo-machete
+# This is due to the `winter-maybe-async` crate not working standalone.
+[package.metadata.cargo-machete]
+ignored = ["async-trait"]
