+++ bin/server/src/server/challenge.rs
