+++ bin/server/src/server/get_note.rs
