+++ bin/server/src/server/get_pow.rs
