+++ bin/server/src/static/index.html
