+++ bin/server/src/static/index.js
