+++ bin/server/src/testing/stub_rpc_api.rs
